+++ src/constants.lua
@@ -1,13 +1,2 @@
 ADDON_NAME = "RaidNotesTBC"
-
--- set to true to enable debug mode
 DEBUG_MODE = false
-
--- set to true to enable test mode, must also enable debug_mode
-TEST_MODE = DEBUG_MODE and false
-TEST_ZONE = "Sunwell Plateau"
-TEST_TARGET = "Lady Sacrolash"
-TEST_INSTANCE_INFO = {
-    { "The Sunwell", 2 },
-    { "Black Temple", 3 }
-}