+++ RELEASE_NOTES.md
@@ -1,9 +1,10 @@
-## 2.0.1 (2022-06-23)
+## 2.0.1 (2022-06-27)
 * [internal] Made code more debuggable
 * [internal] Improved iterating through saved instances
 * [internal] Added test mode and DEBUG_MODE toggle
 * Improved 'Raid Complete' message
 * Updated Notes window to have background and better font size
+* Removed gap before boss notes when no trash notes existed
 
 ## 2.0.0 (2022-06-21)
 * [internal] Overhauled backend